+++ src/main/java/com/gameapi/rha/controller/UserController.java
@@ -78,7 +78,7 @@ public class UserController {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(ACCESS_ERROR);
         }
 
-        User result = UserService.userInfo((String) session.getAttribute("user"));
+        final User result = UserService.userInfo((String) session.getAttribute("user"));
         if (result == null) {
             // Этого быть не может
             return ResponseEntity.status(HttpStatus.OK).body(UNEXPECTED_ERROR);
